added check in view-restaurant to only open controls to owner

+++ src/view-restaurant.php
@@ -66,8 +66,17 @@
 						mysqli_close($con);
 					?>
 					<?php
-						echo '<a href="#"><button class="modify-button" type="submit">Modify</button></a>'
-						echo '<a href="#"><button class="delete-button" type="delete">Delete</button></a>'
+						include_once("./library.php");
+						$con = new mysqli($SERVER, $USERNAME, $PASSWORD, $DATABASE);
+						$rid = $_GET['rid'];
+						$username = $_SESSION['username'];
+						$sql = "SELECT accname FROM `ownedBy` WHERE rid = '" . $rid . "' AND accname = '". $username . "'";
+						$result = $con->query($sql);
+						if ($result->num_rows != 0){
+							echo '<a href="#"><button class="modify-button" type="submit">Modify</button></a>';
+							echo '<a href="#"><button class="delete-button" type="delete">Delete</button></a>';
+						}
+						
 					?>
 				</div>
 			</section>
